+++ test/get_test.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:fanoos_http/fanoos_http.dart';
 import 'package:test/test.dart';
 import 'test.dart';
@@ -13,7 +15,7 @@ void main() {
       'invalid url',
       () async => await httpGet(
           url: 'https://asdasd.com/asdas',
-          onOk: (json) {
+          onOk: (body) {
             shouldThisLineBeCalled(false);
           },
           onError: (e) {
@@ -24,7 +26,8 @@ void main() {
     test('onOk', () async {
       return await httpGet(
           url: 'https://reqres.in/api/users/2',
-          onOk: (json) {
+          bodyParser: jsonDecode,
+          onOk: (body) {
             shouldThisLineBeCalled(true);
           },
           onResponse: (r) {