+++ test/put_test.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:fanoos_http/fanoos_http.dart';
 import 'package:test/test.dart';
 import 'test.dart';
@@ -7,6 +9,7 @@ void main() {
   group('httpPut', () {
     test('simple', () async => httpPut(
       url: 'https://reqres.in/api/users/2',
+      bodyParser: jsonDecode,
       body: {
         'email': 'aligator@test.com',
         'password': 'idontusepassword',