+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.0
+- changed the default `bodyParser` to `plainString`
+- requests will now return `body` if no `onOk` or `onResponse` is provided. 
+
 ## 1.1.0
 - a lot improvements